# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "vibe-quant API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/data/ingest/preview": {
      "post": {
        "tags": [
          "data"
        ],
        "summary": "Ingest Preview",
        "operationId": "ingest_preview_api_data_ingest_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/data/ingest": {
      "post": {
        "tags": [
          "data"
        ],
        "summary": "Start Ingest",
        "operationId": "start_ingest_api_data_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Start Ingest Api Data Ingest Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/data/browse/{symbol}": {
      "get": {
        "tags": [
          "data"
        ],
        "summary": "Browse Data",
        "operationId": "browse_data_api_data_browse__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "interval",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "1m",
              "title": "Interval"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrowseDataResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/data/indicators/{symbol}": {
      "get": {
        "tags": [
          "data"
        ],
        "summary": "Compute Indicators Endpoint",
        "operationId": "compute_indicators_endpoint_api_data_indicators__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "interval",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "1h",
              "title": "Interval"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "name": "indicators",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "[]",
              "title": "Indicators"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndicatorsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/data/quality/{symbol}": {
      "get": {
        "tags": [
          "data"
        ],
        "summary": "Data Quality",
        "description": "Verify the archived 1m klines: exact 1-minute continuity, OHLC sanity,\nflat zero-volume filler runs. Sync endpoint (threadpool): it streams ~1M rows.",
        "operationId": "data_quality_api_data_quality__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataQualityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/data/history": {
      "get": {
        "tags": [
          "data"
        ],
        "summary": "Download History",
        "operationId": "download_history_api_data_history_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Download History Api Data History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BrowseDataResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "interval": {
            "type": "string",
            "title": "Interval"
          },
          "data": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "interval",
          "data"
        ],
        "title": "BrowseDataResponse"
      },
...
      "DataQualityResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "gaps": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Gaps"
          },
          "gap_count": {
            "type": "integer",
            "title": "Gap Count",
            "default": 0
          },
          "missing_bars": {
            "type": "integer",
            "title": "Missing Bars",
            "default": 0
          },
          "zero_volume_runs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Zero Volume Runs",
            "default": []
          },
          "zero_volume_bars": {
            "type": "integer",
            "title": "Zero Volume Bars",
            "default": 0
          },
          "quality_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Score"
          },
          "ohlc_errors": {
            "items": {
              "$ref": "#/components/schemas/OhlcError"
            },
            "type": "array",
            "title": "Ohlc Errors",
            "default": []
          },
          "ohlc_error_count": {
            "type": "integer",
            "title": "Ohlc Error Count",
            "default": 0
          },
          "kline_count": {
            "type": "integer",
            "title": "Kline Count",
            "default": 0
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "gaps",
          "quality_score"
        ],
        "title": "DataQualityResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IndicatorSeries": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "output_name": {
            "type": "string",
            "title": "Output Name"
          },
          "indicator_type": {
            "type": "string",
            "title": "Indicator Type"
          },
          "display_label": {
            "type": "string",
            "title": "Display Label"
          },
          "pane": {
            "type": "string",
            "title": "Pane"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/IndicatorSeriesPoint"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "name",
          "output_name",
          "indicator_type",
          "display_label",
          "pane",
          "params",
          "data"
        ],
        "title": "IndicatorSeries"
      },
...
      "IndicatorsResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "interval": {
            "type": "string",
            "title": "Interval"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/IndicatorSeries"
            },
            "type": "array",
            "title": "Series"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "interval",
          "series"
        ],
        "title": "IndicatorsResponse"
      },
      "IngestPreviewResponse": {
        "properties": {
          "total_months": {
            "type": "integer",
            "title": "Total Months"
          },
          "archived_months": {
            "type": "integer",
            "title": "Archived Months"
          },
          "new_months": {
            "type": "integer",
            "title": "New Months"
          }
        },
        "type": "object",
        "required": [
          "total_months",
          "archived_months",
          "new_months"
        ],
        "title": "IngestPreviewResponse"
      },
      "IngestRequest": {
        "properties": {
          "symbols": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Symbols"
          },
          "start_date": {
            "type": "string",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "title": "End Date"
          },
          "interval": {
            "type": "string",
            "title": "Interval",
            "default": "1m"
          }
        },
        "type": "object",
        "required": [
          "symbols",
          "start_date",
          "end_date"
        ],
        "title": "IngestRequest"
      },
...
      "OhlcError": {
        "properties": {
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "error_type": {
            "type": "string",
            "title": "Error Type"
          },
          "values": {
            "additionalProperties": true,
            "type": "object",
            "title": "Values"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "error_type",
          "values"
        ],
        "title": "OhlcError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}