# vibe_quant/api/routers/data.py
"""Data management router."""

from __future__ import annotations

import contextlib
import logging
import os
import sqlite3
import sys
import time
from datetime import UTC, datetime
from typing import TYPE_CHECKING, Annotated

if TYPE_CHECKING:
    from vibe_quant.data.archive import RawDataArchive

from fastapi import APIRouter, Depends, HTTPException, Query

from vibe_quant.api.deps import get_catalog_manager, get_job_manager
from vibe_quant.api.schemas.data import (
    BrowseDataResponse,
    DataCoverageItem,
    DataCoverageResponse,
    DataQualityResponse,
    DataStatusResponse,
    IndicatorsResponse,
    IngestPreviewResponse,
    IngestRequest,
    OhlcError,
)
from vibe_quant.data.catalog import CatalogManager
from vibe_quant.jobs.manager import BacktestJobManager

logger = logging.getLogger(__name__)

router = APIRouter(prefix="/api/data", tags=["data"])

_INTERVAL_MINUTES = {"1m": 1, "5m": 5, "15m": 15, "30m": 30, "1h": 60, "4h": 240, "1d": 1440}


def _next_data_run_id() -> int:
    """Generate a unique negative run_id for data jobs.

    Uses negative timestamp-based IDs to avoid collision with positive
    backtest run IDs. Microsecond precision prevents duplicates under
    normal concurrency.
    """
    return -int(time.time() * 1_000_000)


def _parse_interval_minutes(interval: str) -> int:
    """Convert interval string to minutes."""
    if interval in _INTERVAL_MINUTES:
        return _INTERVAL_MINUTES[interval]
    return 1


CatMgr = Annotated[CatalogManager, Depends(get_catalog_manager)]
JobMgr = Annotated[BacktestJobManager, Depends(get_job_manager)]


def _get_archive() -> RawDataArchive:
    from vibe_quant.data.archive import RawDataArchive

    return RawDataArchive()


def _dir_size(path: str) -> int:
    total = 0
    for dirpath, _dirnames, filenames in os.walk(path):
        for f in filenames:
            fp = os.path.join(dirpath, f)
            with contextlib.suppress(OSError):
                total += os.path.getsize(fp)
    return total


# --- Storage status ---


@router.get("/status", response_model=DataStatusResponse)
async def data_status() -> DataStatusResponse:
    from vibe_quant.data.archive import DEFAULT_ARCHIVE_PATH
    from vibe_quant.data.catalog import DEFAULT_CATALOG_PATH

    archive_size = 0
    if DEFAULT_ARCHIVE_PATH.exists():
        archive_size = DEFAULT_ARCHIVE_PATH.stat().st_size

    catalog_size = _dir_size(str(DEFAULT_CATALOG_PATH)) if DEFAULT_CATALOG_PATH.exists() else 0

    return DataStatusResponse(
        archive_size_bytes=archive_size,
        catalog_size_bytes=catalog_size,
        total_size_bytes=archive_size + catalog_size,
    )


# --- Coverage ---


@router.get("/coverage", response_model=DataCoverageResponse)
async def data_coverage(catalog: CatMgr) -> DataCoverageResponse:
    archive = _get_archive()
    try:
        symbols = archive.get_symbols()
        items: list[DataCoverageItem] = []

        for symbol in symbols:
            kline_count = archive.get_kline_count(symbol, "1m")
            date_range = archive.get_date_range(symbol, "1m")

            start_date = ""
            end_date = ""
            if date_range:
                start_date = datetime.fromtimestamp(date_range[0] / 1000, tz=UTC).strftime(
                    "%Y-%m-%d"
                )
                end_date = datetime.fromtimestamp(date_range[1] / 1000, tz=UTC).strftime("%Y-%m-%d")

            bar_count = catalog.get_bar_count(symbol, "1m")

            # Count funding rates via direct query
            row = archive.conn.execute(
                "SELECT COUNT(*) FROM raw_funding_rates WHERE symbol = ?",
                (symbol,),
            ).fetchone()
            funding_rate_count = row[0] if row else 0

            items.append(
                DataCoverageItem(
                    symbol=symbol,
                    start_date=start_date,
                    end_date=end_date,
                    kline_count=kline_count,
                    bar_count=bar_count,
                    funding_rate_count=funding_rate_count,
                )
            )

        return DataCoverageResponse(coverage=items)
    finally:
        archive.close()


# --- Symbols ---


@router.get("/symbols")
async def list_symbols() -> list[str]:
    archive = _get_archive()
    try:
        return archive.get_symbols()
    finally:
        archive.close()


# --- Ingest preview ---


@router.post("/ingest/preview", response_model=IngestPreviewResponse)
async def ingest_preview(body: IngestRequest) -> IngestPreviewResponse:
    from vibe_quant.data.ingest import get_download_preview

    try:
        start_dt = datetime.strptime(body.start_date, "%Y-%m-%d").replace(tzinfo=UTC)
        end_dt = datetime.strptime(body.end_date, "%Y-%m-%d").replace(tzinfo=UTC)
    except ValueError as exc:
        raise HTTPException(status_code=400, detail=f"Invalid date format: {exc}") from exc

    preview = get_download_preview(body.symbols, start_dt, end_dt)
    total = len(preview)
    archived = sum(1 for p in preview if p["Status"] == "Archived")
    new = total - archived

    return IngestPreviewResponse(
        total_months=total,
        archived_months=archived,
        new_months=new,
    )


# --- Ingest (background job) ---


def _start_data_job(
    jobs: BacktestJobManager, job_type: str, log_prefix: str, data_args: list[str]
) -> tuple[int, int]:
    """Spawn a data CLI command via the heartbeating job wrapper.

    Returns (job_id, pid). ``job_id`` (negative, never a backtest run id)
    keys the job row and the ``/api/data/ingest/{job_id}/progress`` stream.
    """
    job_id = _next_data_run_id()
    log_file = f"logs/{log_prefix}_{datetime.now(UTC).strftime('%Y%m%d_%H%M%S')}.log"
    command = [
        sys.executable,
        "-m",
        "vibe_quant.jobs.data_job",
        "--run-id",
        str(job_id),
        "--db",
        str(jobs.db_path),
        "--",
        *data_args,
    ]
    try:
        pid = jobs.start_job(job_id, job_type, command, log_file=log_file)
    except ValueError as exc:
        raise HTTPException(status_code=409, detail=str(exc)) from exc
    return job_id, pid


@router.post("/ingest", status_code=202)
async def start_ingest(body: IngestRequest, jobs: JobMgr) -> dict[str, object]:
    job_id, pid = _start_data_job(
        jobs,
        "data_ingest",
        "ingest",
        [
            "ingest",
            "--symbols",
            ",".join(body.symbols),
            "--start",
            body.start_date,
            "--end",
            body.end_date,
        ],
    )
    logger.info("data ingest started job_id=%d pid=%d symbols=%s", job_id, pid, body.symbols)
    return {
        "status": "started",
        "job_id": job_id,
        "pid": pid,
        "symbols": body.symbols,
        "start_date": body.start_date,
        "end_date": body.end_date,
    }


# --- Update (background job) ---


@router.post("/update", status_code=202)
async def start_update(jobs: JobMgr) -> dict[str, object]:
    job_id, pid = _start_data_job(jobs, "data_update", "update", ["update"])
    logger.info("data update started job_id=%d pid=%d", job_id, pid)
    return {"status": "started", "job_id": job_id, "pid": pid}


# --- Rebuild catalog ---


@router.post("/rebuild", status_code=202)
async def rebuild_catalog(jobs: JobMgr) -> dict[str, object]:
    job_id, pid = _start_data_job(
        jobs, "catalog_rebuild", "rebuild", ["rebuild", "--from-archive"]
    )
    logger.info("catalog rebuild started job_id=%d pid=%d", job_id, pid)
    return {"status": "started", "job_id": job_id, "pid": pid}


# --- Browse OHLCV data ---


@router.get("/browse/{symbol}", response_model=BrowseDataResponse)
async def browse_data(
    symbol: str,
    interval: str = Query(default="1m"),
    start: str | None = Query(default=None),
    end: str | None = Query(default=None),
) -> BrowseDataResponse:
    archive = _get_archive()
    try:
        start_ts: int | None = None
        end_ts: int | None = None

        if start:
            try:
                start_ts = int(
                    datetime.strptime(start, "%Y-%m-%d").replace(tzinfo=UTC).timestamp() * 1000
                )
            except ValueError as exc:
                raise HTTPException(status_code=400, detail=f"Invalid start date: {exc}") from exc
        if end:
            try:
                end_ts = int(
                    datetime.strptime(end, "%Y-%m-%d").replace(tzinfo=UTC).timestamp() * 1000
                )
            except ValueError as exc:
                raise HTTPException(status_code=400, detail=f"Invalid end date: {exc}") from exc

        if interval not in _INTERVAL_MINUTES:
            raise HTTPException(
                status_code=400,
                detail=f"Invalid interval '{interval}'. Must be one of: {', '.join(_INTERVAL_MINUTES)}",
            )
        interval_minutes = _INTERVAL_MINUTES[interval]
        bucket_ms = interval_minutes * 60 * 1000

        # Build WHERE clause
        conditions = ["symbol = ?", "interval = '1m'"]
        params: list[object] = [symbol]
        if start_ts is not None:
            conditions.append("open_time >= ?")
            params.append(start_ts)
        if end_ts is not None:
            conditions.append("open_time < ?")
            params.append(end_ts)
        where = " AND ".join(conditions)

        # Single ordered scan, resample in Python (cursor streaming avoids full materialization)
        sql = f"""
            SELECT open_time, open, high, low, close, volume, close_time
            FROM raw_klines
            WHERE {where}
            ORDER BY open_time
        """
        data: list[dict[str, object]] = []

        if interval_minutes <= 1:
            for r in archive.conn.execute(sql, tuple(params)):
                data.append(
                    {
                        "open_time": r[0],
                        "open": r[1],
                        "high": r[2],
                        "low": r[3],
                        "close": r[4],
                        "volume": r[5],
                        "close_time": r[6],
                    }
                )
        else:
            bucket: dict[str, object] | None = None
            for r in archive.conn.execute(sql, tuple(params)):
                ot = r[0]
                bucket_start = (ot // bucket_ms) * bucket_ms
                if bucket is None or bucket["open_time"] != bucket_start:
                    if bucket is not None:
                        data.append(bucket)
                    bucket = {
                        "open_time": bucket_start,
                        "open": r[1],
                        "high": r[2],
                        "low": r[3],
                        "close": r[4],
                        "volume": r[5],
                        "close_time": bucket_start + bucket_ms - 1,
                    }
                else:
                    if r[2] > bucket["high"]:  # type: ignore[operator]
                        bucket["high"] = r[2]
                    if r[3] < bucket["low"]:  # type: ignore[operator]
                        bucket["low"] = r[3]
                    bucket["close"] = r[4]
                    bucket["volume"] = bucket["volume"] + r[5]  # type: ignore[operator]
                    bucket["close_time"] = r[6]
            if bucket is not None:
                data.append(bucket)

        return BrowseDataResponse(symbol=symbol, interval=interval, data=data)
    finally:
        archive.close()


# --- Indicator computation ---


@router.get("/indicators/{symbol}", response_model=IndicatorsResponse)
async def compute_indicators_endpoint(
    symbol: str,
    interval: str = Query(default="1h"),
    start: str | None = Query(default=None),
    end: str | None = Query(default=None),
    indicators: str = Query(default="[]"),
) -> IndicatorsResponse:
    import json

    import pandas as pd

    from vibe_quant.data.indicator_compute import compute_indicators

    # Parse indicators JSON
    try:
        indicator_configs: list[dict[str, object]] = json.loads(indicators)
    except json.JSONDecodeError as exc:
        raise HTTPException(status_code=400, detail=f"Invalid indicators JSON: {exc}") from exc

    if not indicator_configs:
        return IndicatorsResponse(symbol=symbol, interval=interval, series=[])

    if interval not in _INTERVAL_MINUTES:
        raise HTTPException(
            status_code=400,
            detail=f"Invalid interval '{interval}'. Must be one of: {', '.join(_INTERVAL_MINUTES)}",
        )

    archive = _get_archive()
    try:
        start_ts: int | None = None
        end_ts: int | None = None
        if start:
            try:
                start_ts = int(
                    datetime.strptime(start, "%Y-%m-%d").replace(tzinfo=UTC).timestamp() * 1000
                )
            except ValueError as exc:
                raise HTTPException(status_code=400, detail=f"Invalid start date: {exc}") from exc
        if end:
            try:
                end_ts = int(
                    datetime.strptime(end, "%Y-%m-%d").replace(tzinfo=UTC).timestamp() * 1000
                )
            except ValueError as exc:
                raise HTTPException(status_code=400, detail=f"Invalid end date: {exc}") from exc

        interval_minutes = _INTERVAL_MINUTES[interval]
        bucket_ms = interval_minutes * 60 * 1000

        # Fetch 1m candles
        conditions = ["symbol = ?", "interval = '1m'"]
        params: list[object] = [symbol]
        if start_ts is not None:
            conditions.append("open_time >= ?")
            params.append(start_ts)
        if end_ts is not None:
            conditions.append("open_time < ?")
            params.append(end_ts)
        where = " AND ".join(conditions)

        sql = f"""
            SELECT open_time, open, high, low, close, volume
            FROM raw_klines
            WHERE {where}
            ORDER BY open_time
        """
        rows = archive.conn.execute(sql, tuple(params)).fetchall()

        if not rows:
            return IndicatorsResponse(symbol=symbol, interval=interval, series=[])

        # Build DataFrame
        raw_df = pd.DataFrame(rows, columns=["open_time", "open", "high", "low", "close", "volume"])

        # Aggregate to requested interval if > 1m
        if interval_minutes > 1:
            raw_df["bucket"] = (raw_df["open_time"] // bucket_ms) * bucket_ms
            df = raw_df.groupby("bucket").agg(
                open_time=("bucket", "first"),
                open=("open", "first"),
                high=("high", "max"),
                low=("low", "min"),
                close=("close", "last"),
                volume=("volume", "sum"),
            ).reset_index(drop=True)
        else:
            df = raw_df

        series = compute_indicators(df, indicator_configs)
        return IndicatorsResponse(symbol=symbol, interval=interval, series=series)
    finally:
        archive.close()


# --- Data quality ---


def _iso_ms(ts: int) -> str:
    return datetime.fromtimestamp(ts / 1000, tz=UTC).strftime("%Y-%m-%d %H:%M")


@router.get("/quality/{symbol}", response_model=DataQualityResponse)
def data_quality(symbol: str) -> DataQualityResponse:
    """Verify the archived 1m klines: exact 1-minute continuity, OHLC sanity,
    flat zero-volume filler runs. Sync endpoint (threadpool): it streams ~1M rows.
    """
    from vibe_quant.data.verify import scan_klines

    archive = _get_archive()
    try:
        cursor = archive.conn.execute(
            """SELECT open_time, open, high, low, close, volume FROM raw_klines
               WHERE symbol = ? AND interval = '1m' ORDER BY open_time""",
            (symbol,),
        )
        cursor.row_factory = sqlite3.Row
        result = scan_klines(cursor)
    finally:
        archive.close()

    if result["kline_count"] == 0:
        return DataQualityResponse(
            symbol=symbol,
            gaps=[],
            quality_score=None,
            error=f"No 1m data archived for {symbol}",
        )

    gaps: list[dict[str, object]] = [
        {"start": _iso_ms(a), "end": _iso_ms(b), "missing_bars": max(step - 1, 0)}
        for a, b, step in result["gaps"]
    ]
    filler: list[dict[str, object]] = [
        {"start": _iso_ms(a), "end": _iso_ms(b), "bars": n}
        for a, b, n in result["zero_volume_runs"]
    ]
    ohlc_errors = [
        OhlcError(timestamp=_iso_ms(ts), error_type="ohlc_inconsistent", values={"message": msg})
        for ts, msg in result["ohlc_errors"]
    ]
    missing = sum(max(step - 1, 0) for _a, _b, step in result["gaps"])
    filler_bars = sum(n for _a, _b, n in result["zero_volume_runs"])
    bad_ohlc_bars = len({ts for ts, _msg in result["ohlc_errors"]})
    expected = result["kline_count"] + missing
    quality_score = max(0.0, 1.0 - (missing + filler_bars + bad_ohlc_bars) / expected)

    return DataQualityResponse(
        symbol=symbol,
        gaps=gaps[:500],
        gap_count=len(gaps),
        missing_bars=missing,
        zero_volume_runs=filler[:500],
        zero_volume_bars=filler_bars,
        quality_score=quality_score,
        ohlc_errors=ohlc_errors[:50],  # Cap to 50 for response size
        ohlc_error_count=len(ohlc_errors),
        kline_count=result["kline_count"],
    )


# --- Download history ---


@router.get("/history")
async def download_history(limit: int = Query(default=50, ge=1, le=200)) -> list[dict[str, object]]:
    archive = _get_archive()
    try:
        rows = archive.conn.execute(
            """
            SELECT
                symbol,
                interval,
                MIN(open_time) AS min_time,
                MAX(open_time) AS max_time,
                COUNT(*) AS row_count,
                MAX(downloaded_at) AS last_downloaded
            FROM raw_klines
            GROUP BY symbol, interval
            ORDER BY MAX(downloaded_at) DESC
            LIMIT ?
            """,
            (limit,),
        ).fetchall()
        result: list[dict[str, object]] = []
        for r in rows:
            min_ts = r[2]
            max_ts = r[3]
            start_date = (
                datetime.fromtimestamp(min_ts / 1000, tz=UTC).strftime("%Y-%m-%d")
                if min_ts
                else None
            )
            end_date = (
                datetime.fromtimestamp(max_ts / 1000, tz=UTC).strftime("%Y-%m-%d")
                if max_ts
                else None
            )
            result.append(
                {
                    "symbol": r[0],
                    "interval": r[1],
                    "start_date": start_date,
                    "end_date": end_date,
                    "rows": r[4],
                    "timestamp": r[5],
                }
            )
        return result
    finally:
        archive.close()